moves asynch request to request file
