+++ lib/books.js
@@ -1,11 +1,11 @@
 module.exports = {makeCall}
+requests = require('./requests')
 
 const request = require('request');
 const tableHeaders = '<thead><tr><th scope="col">Price (BTC)</th>'
                     + '<th scope="col">Bittrex</th>'
                     + '<th scope="col">Poloniex</th>'
                     + '<th scope="col">Total</th>'
-                    // + '<th scope="col">Sum</th></tr></thead>'
 
 
 let url1 = "https://bittrex.com/api/v1.1/public/getorderbook?market=BTC-ETH&type=both"
@@ -13,22 +13,12 @@ let url2 = "https://poloniex.com/public?command=returnOrderBook&currencyPair=BTC
 
 let data = ["", ""]
 
-function requestAsync(url) {
-    return new Promise(function(resolve, reject) {
-        request(url, function(err, res, body) {
-            if (err) { return reject(err); }
-            return resolve(JSON.parse(body));
-        });
-    });
-}
-
 function makeCall(){
-  Promise.all([requestAsync(url1), requestAsync(url2)])
+  Promise.all([requests.requestAsync(url1), requests.requestAsync(url2)])
       .then(function(allData) {
         var lowerCaseData = JSON.stringify(allData).toLowerCase()
         var formattedData = formatData(JSON.parse(lowerCaseData))
         data = generateOrderBooks(formattedData)
-        // data = ["Bittrex: " + formattedData[0].bids[0].rate, "Poloniex: " + formattedData[1].bids[0].rate]
       });
       return JSON.stringify(data)
 }
