+++ lib/requests.js
@@ -1,0 +1,12 @@
+module.exports = {requestAsync}
+const request = require('request');
+
+
+function requestAsync(url) {
+    return new Promise(function(resolve, reject) {
+        request(url, function(err, res, body) {
+            if (err) { return reject(err); }
+            return resolve(JSON.parse(body));
+        });
+    });
+}
